Changes to fix compatibility with command-line builds (using xcodebuild).
The plug-in was set to require GC and it seems xcodebuild does not use it, which caused a bundle load error to be displayed when building from the command-line.
I’ve also removed a message which was displayed when swizzling of -keyDown: failed, as the textview class isn’t loaded by xcodebuild so it would always be shown.